feat/training_complete
