+++ src/config/config.yaml
@@ -3,9 +3,9 @@ project:
 
 model:
   img_size: 640
-  batch_size: 16
-  nb_epochs: 10
-  device: "cpu"
+  batch_size: 8
+  nb_epochs: 1
+  device: 1
 
 evaluation:
   key: value
